+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed download button in the File Manager (#121).
+- Fixed ordering of study collections in the 'Add Study to Collection' dialog (#125).
 
 ## [0.9.4] - 2023-10-20